requirements.txt
appdirs==1.4.4
attrs==20.3.0
behave==1.2.6
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
dnslib==0.9.14
dnspython==1.16.0
docopt==0.6.2
filelock==3.0.12
flake8==3.8.4
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
localstack==0.12.2
localstack-client==1.7
localstack-ext==0.12.0
mccabe==0.6.1
packaging==20.4
parse==1.18.0
parse-type==0.5.2
pluggy==0.13.1
py==1.9.0
pyaes==1.6.1
pycodestyle==2.6.0
pycryptodome==3.9.9
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.2
tox==3.20.1
urllib3==1.26.2
virtualenv==20.2.1
